rag_auth_lib/examples/llama_index_examples/pyproject.toml
[tool.poetry]
name = "llama_index_examples"
version = "0.1.0"
description = "Pangea Generalized Auth library for Llama Index"
authors = ["Apurv Jawle <apurv.jawle@pangea.cloud>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pangea-sdk = "^5.1.0"
google-auth-oauthlib = "^1.2.1"
llama-index = "0.11.11"
llama-index-llms-bedrock = "0.2.1"
llama-index-embeddings-bedrock = "0.3.1"
llama-index-readers-google = "^0.4.1"
llama-index-readers-confluence = "^0.2.1"
llama-index-readers-jira = "^0.3.1"
pangea-auth-llama-index = { path = "../../pangea-auth-llama-index", develop = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
